do some renaming
