+++ src/easydynamics/sample/sample_model.py
